+++ R/revdep.R
@@ -117,11 +117,11 @@ revdep_check <- function(pkg = ".", recursive = FALSE, ignore = NULL,
   if (!file.exists(srcpath))
     dir.create(srcpath)
 
-  message("Installing ", pkg$package, " in version ", pkg$version, " from ", pkg$path)
+  message("Installing ", pkg$package, " ", pkg$version, " from ", pkg$path)
   withr::with_libpaths(libpath, install(pkg, reload = FALSE, quiet = TRUE))
   on.exit(remove.packages(pkg$package, libpath), add = TRUE)
 
-  old <- withr::with_envvar(c(
+  withr::with_envvar(c(
     NOT_CRAN = "false",
     RGL_USE_NULL = "true"
   ), {